requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aiofiles
aiohttp
aiosignal
anyio
async-timeout
attrs
bitarray
blinker
cachetools
certifi
charset-normalizer
ci-info
click
colorama
configobj
configparser
Deprecated
elastic-transport
elasticsearch~=8.13.0
etelemetry
exceptiongroup
filelock
Flask~=2.3.3
Flask-Cors
frontend
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
h11
httplib2
idna
importlib-metadata
isodate
itsdangerous
Jinja2
looseversion
lxml
MarkupSafe
multidict
networkx
nibabel
numpy
oauthlib
openai
packaging
pandas
pathlib
pdfreader
pikepdf
pillow
protobuf
prov
pyasn1
pyasn1-modules
pycryptodome
pydot
PyMuPDF
PyMuPDFb
pyparsing
pypdf~=4.1.0
PyPDF2
python-dateutil
pytz
pyxnat
rdflib
regex
requests
requests-oauthlib
rsa
scipy
simplejson
six
sniffio
starlette
tiktoken
tqdm
traits
tzdata
uritemplate
urllib3
uvicorn
Werkzeug
wrapt
yarl
zipp
langchain~=0.1.14
langchain_elasticsearch
slackeventsapi~=3.0.1
nltk~=3.8.1
python-dotenv~=1.0.1
slackclient
langchain_experimental
presidio_analyzer
presidio-anonymizer
Faker
spacy[transformers]
torch